switch encoder to use quadrature_encoder.pio
This runs the quadrature_encoder and the button debouncer on different
PIO instances.

This is required so that the slow clock of the debounce doesn't mess up
the quadrature decoder.

The quadrature program takes up almost all the instruction memory of the
PIO instance it's running on, so there's not room to add a bunch of NOPs
to the debouncer to slow it down.